requirements.txt
python-dotenv
Pillow
requests
django-crispy-forms
